% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-arrange.R
\name{arrange.substrait_dplyr_query}
\alias{arrange.substrait_dplyr_query}
\title{Arrange rows by columns values in a Substrait query}
\usage{
\method{arrange}{substrait_dplyr_query}(.data, ..., .by_group = FALSE)
}
\arguments{
\item{.data}{A \code{\link[=substrait_dplyr_query]{substrait_dplyr_query()}}}

\item{...}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Variables, or functions of
variables. Use \code{\link[dplyr:desc]{desc()}} to sort a variable in descending order.}

\item{.by_group}{If \code{TRUE}, will sort first by grouping variable. Applies to
grouped data frames only.}
}
\value{
A \code{\link[=substrait_dplyr_query]{substrait_dplyr_query()}}
}
\description{
Arrange rows by columns values in a Substrait query
}
\examples{
dplyr::arrange(
  substrait_dplyr_query(mtcars),
  hp, carb
)
}
