% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-arrow.R
\name{arrow_substrait_compiler}
\alias{arrow_substrait_compiler}
\title{Create an Arrow Substrait Compiler}
\usage{
arrow_substrait_compiler(object, ...)
}
\arguments{
\item{object}{A \code{\link[=data.frame]{data.frame()}}, \link[arrow:Table]{arrow::Table}, \link[arrow:RecordBatch]{arrow::RecordBatch},
or \link[arrow:Dataset]{arrow::Dataset}, or anything else that can be written to a
parquet file using \code{\link[arrow:write_parquet]{arrow::write_parquet()}}.}

\item{...}{Unused.}
}
\value{
A \link{SubstraitCompiler} subclass that will
}
\description{
Create an Arrow Substrait Compiler
}
