% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrait-proto.R
\name{as_substrait}
\alias{as_substrait}
\alias{from_substrait}
\alias{as_substrait_expression}
\alias{as_substrait_type}
\title{Convert to and from 'Substrait' messages}
\usage{
as_substrait(x, .ptype = NULL, ...)

from_substrait(msg, x, ...)

as_substrait_expression(x, ...)

as_substrait_type(x, ...)
}
\arguments{
\item{x}{An object to convert to or from a 'Substrait' message.
Note that both \code{as_substrait()} and \code{from_substrait()} dispatch
on \code{x}.}

\item{.ptype}{A string of the \code{.qualified_name} or a prototype message
of the correct type.}

\item{...}{Passed to S3 methods}

\item{msg}{A substrait message (e.g., created using \code{\link[=substrait_create]{substrait_create()}}).}
}
\value{
An RProtoBuf::Message or substrait_proto_message (e.g.,
created by \code{\link[=substrait_create]{substrait_create()}})
}
\description{
Convert to and from 'Substrait' messages
}
\examples{
as_substrait(substrait$Type$Boolean$create(type_variation_reference = 1))
as_substrait_expression(4L)
as_substrait_expression(3.14)

}
