% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-select.R
\name{as_substrait.substrait_op}
\alias{as_substrait.substrait_op}
\title{Create a substrait plan from}
\usage{
\method{as_substrait}{substrait_op}(x, .ptype = NULL, ...)
}
\arguments{
\item{x}{A substrait_op object}

\item{.ptype}{A string of the \code{.qualified_name} or a prototype message
of the correct type.}

\item{...}{Passed to S3 methods}
}
\value{
A substrait plan
}
\description{
Create a substrait plan from
}
