% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-projections.R
\name{build_projections}
\alias{build_projections}
\title{Take selected columns and create the appropriate substrait message}
\usage{
build_projections(df, projections)
}
\arguments{
\item{df}{data}

\item{projections}{vector of selected columns}
}
\description{
Take selected columns and create the appropriate substrait message
}
