% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-select.R
\name{build_substrait}
\alias{build_substrait}
\alias{build_substrait.substrait_select}
\title{Build a substrait expression from a dplyr select object}
\usage{
build_substrait(x)

\method{build_substrait}{substrait_select}(x)
}
\arguments{
\item{x}{Object of class \code{substrait_select}}
}
\value{
List of selection expressions
}
\description{
Build a substrait expression from a dplyr select object
}
