% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-dplyr.R
\name{count.SubstraitCompiler}
\alias{count.SubstraitCompiler}
\title{rdname select.SubstraitCompiler}
\usage{
\method{count}{SubstraitCompiler}(.data, ..., wt = NULL, sort = FALSE, name = NULL)
}
\description{
rdname select.SubstraitCompiler
}
