% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-dplyr.R
\name{distinct.SubstraitCompiler}
\alias{distinct.SubstraitCompiler}
\title{rdname select.SubstraitCompiler}
\usage{
\method{distinct}{SubstraitCompiler}(.data, ..., .keep_all = FALSE)
}
\description{
rdname select.SubstraitCompiler
}
