% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-duckdb.R
\name{duckdb_substrait_compiler}
\alias{duckdb_substrait_compiler}
\title{Create an DuckDB Substrait Compiler}
\usage{
duckdb_substrait_compiler(object, ...)
}
\arguments{
\item{object}{A \code{\link[=data.frame]{data.frame()}}, \link[arrow:Table]{arrow::Table}, \link[arrow:RecordBatch]{arrow::RecordBatch},
or \link[arrow:Dataset]{arrow::Dataset}, or anything else that can be written to a
parquet file using \code{\link[arrow:write_parquet]{arrow::write_parquet()}}.}

\item{...}{Unused.}
}
\value{
A \link{SubstraitCompiler} subclass.
}
\description{
Create an DuckDB Substrait Compiler
}
