% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{register_functions}
\alias{register_functions}
\alias{register_functions_yaml}
\alias{default_function_registry}
\alias{new_function_registry}
\title{Register functions}
\usage{
register_functions(
  scalar = list(),
  aggregate = list(),
  window = list(),
  registry = default_function_registry()
)

register_functions_yaml(path, registry = default_function_registry())

default_function_registry()

new_function_registry()
}
\arguments{
\item{scalar}{A named list of scalar function definitions}

\item{aggregate}{A named list of aggregate function definitions}

\item{window}{A list of window function definitions}

\item{registry}{The \code{\link[=default_function_registry]{default_function_registry()}} or a custom registry
created using \code{\link[=new_function_registry]{new_function_registry()}}.}

\item{path}{A path to .yaml function definitions.}
}
\value{
\itemize{
\item \code{register_function()} and \code{register_functions_yaml()} return \code{registry},
(invisibly)
\item \code{default_function_registry()} and \code{new_function_registry()} return a
\code{list()} with items "scalar", "window", and "aggregate".
}
}
\description{
Register functions
}
\examples{
default_function_registry()
(registry <- new_function_registry())

}
