% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{resolve_function_by_name}
\alias{resolve_function_by_name}
\title{Resolve function references}
\usage{
resolve_function_by_name(
  name,
  args,
  registry = default_function_registry(),
  type = "scalar"
)
}
\arguments{
\item{name}{The function name}

\item{args}{A vector of (possibly named) arguments used to resolve
an implementation.}

\item{registry}{The \code{\link[=default_function_registry]{default_function_registry()}} or a custom registry
created using \code{\link[=new_function_registry]{new_function_registry()}}.}

\item{type}{One of "scalar", "aggregate" or "windows"}
}
\value{
A Substrait message object of type Expression.ScalarFunction,
Expression.WindowFunction, or AggregateFunction.
}
\description{
Resolve function references
}
