% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-dplyr.R
\name{select.substrait_compiler}
\alias{select.substrait_compiler}
\alias{rename.substrait_compiler}
\alias{filter.substrait_compiler}
\alias{mutate.substrait_compiler}
\alias{transmute.substrait_compiler}
\alias{arrange.substrait_compiler}
\alias{collect.substrait_compiler}
\title{dplyr verb implementations}
\usage{
\method{select}{substrait_compiler}(.data, ...)

\method{rename}{substrait_compiler}(.data, ...)

\method{filter}{substrait_compiler}(.data, ...)

\method{mutate}{substrait_compiler}(.data, ...)

\method{transmute}{substrait_compiler}(.data, ...)

\method{arrange}{substrait_compiler}(.data, ...)

\method{collect}{substrait_compiler}(x, ...)
}
\arguments{
\item{.data, x}{A \code{\link[=substrait_compiler]{substrait_compiler()}}}

\item{...}{\itemize{
\item \code{select()}: see \code{\link[dplyr:select]{dplyr::select()}}
\itemize{
\item \code{rename()}: see \code{\link[dplyr:rename]{dplyr::rename()}}
\item \code{filter()}: see \code{\link[dplyr:filter]{dplyr::filter()}}
\item \code{mutate()}: see \code{\link[dplyr:mutate]{dplyr::mutate()}}
\item \code{arrange()}: see \code{\link[dplyr:arrange]{dplyr::arrange()}}
}
}}
}
\value{
A modified \code{\link[=substrait_compiler]{substrait_compiler()}}
}
\description{
dplyr verb implementations
}
\examples{
\dontshow{if (requireNamespace("dplyr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
compiler <- substrait_compiler(mtcars)
dplyr::select(compiler, mpg2 = mpg)
dplyr::rename(compiler, mpg2 = mpg)
dplyr::filter(compiler, mpg > 20)
dplyr::mutate(compiler, mpg + 10)
dplyr::transmute(compiler, mpg + 10)
dplyr::arrange(compiler, desc(mpg))
\dontshow{\}) # examplesIf}
}
