% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-select.R
\name{select.substrait_op}
\alias{select.substrait_op}
\title{Select column}
\usage{
\method{select}{substrait_op}(.data, ...)
}
\arguments{
\item{.data}{substrait_Rel object}
}
\description{
Select column
}
