% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrait-package.R
\docType{package}
\name{substrait-package}
\title{substrait: 'Substrait' Cross-Language Serialization for Relational Algebra}
\description{
Provides an R interface to the 'Substrait' cross-language serialization for relational algebra.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

Authors:
\itemize{
  \item Nic Crane \email{thisisnic@gmail.com}
  \item Jonathan Keane \email{jkeane@gmail.com}
}

}
\keyword{internal}
