% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{substrait-type}
\alias{substrait-type}
\alias{substrait_boolean}
\alias{substrait_i8}
\alias{substrait_i16}
\alias{substrait_i32}
\alias{substrait_i64}
\alias{substrait_fp32}
\alias{substrait_fp64}
\alias{substrait_string}
\alias{substrait_binary}
\alias{substrait_timestamp}
\alias{substrait_timestamp_tz}
\alias{substrait_date}
\alias{substrait_time}
\alias{substrait_interval_year}
\alias{substrait_interval_day}
\alias{substrait_uuid}
\title{Substrait types}
\usage{
substrait_boolean()

substrait_i8()

substrait_i16()

substrait_i32()

substrait_i64()

substrait_fp32()

substrait_fp64()

substrait_string()

substrait_binary()

substrait_timestamp()

substrait_timestamp_tz()

substrait_date()

substrait_time()

substrait_interval_year()

substrait_interval_day()

substrait_uuid()
}
\value{
A substrait.Type proto object
}
\description{
Substrait types
}
\examples{
substrait_boolean()
substrait_i32()
substrait_fp64()
substrait_string()

}
