% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-generated.R
\docType{data}
\name{substrait}
\alias{substrait}
\title{Create raw Substrait types}
\format{
An object of class \code{list} of length 27.
}
\usage{
substrait
}
\description{
Create raw Substrait types
}
\keyword{datasets}
