% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate-rel.R
\name{substrait_aggregate}
\alias{substrait_aggregate}
\alias{substrait_group_by}
\alias{substrait_ungroup}
\title{Aggregate}
\usage{
substrait_aggregate(.compiler, ...)

substrait_group_by(.compiler, ...)

substrait_ungroup(.compiler)
}
\arguments{
\item{.compiler}{A \code{\link[=substrait_compiler]{substrait_compiler()}} or object that can be coerced to one}

\item{...}{\itemize{
\item \code{substrait_aggregate()}: A named list of expressions to be evaluated in the context
of the aggregation.
\itemize{
\item \code{substrait_group_by()}: A list of expressions to be used as groupings
for aggregation.
}
}}
}
\value{
A modified \code{.compiler}.
}
\description{
Aggregate
}
