% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-arrow.R
\name{substrait_colnames}
\alias{substrait_colnames}
\title{Get column names for a relation}
\usage{
substrait_colnames(x)
}
\arguments{
\item{x}{A relation}
}
\value{
A vector of column names
}
\description{
Get column names for a relation
}
