% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrait-proto.R
\name{substrait_create}
\alias{substrait_create}
\title{Create 'Substrait' objects}
\usage{
substrait_create(.qualified_name, ...)
}
\arguments{
\item{.qualified_name}{The fully qualified name of the message type
or enum (e.g., "substrait.Type.Boolean")}

\item{...}{Arguments passed to the constructor.}
}
\value{
An object of class "substrait_proto".
}
\description{
Create 'Substrait' objects
}
\examples{
substrait_create("substrait.Type.Boolean", 1, 2)
substrait$Type$Boolean$create(1, 2)

}
