% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrait_dplyr_query.R
\name{substrait_dplyr_query}
\alias{substrait_dplyr_query}
\title{Create a substrait_dplyr_query object}
\usage{
substrait_dplyr_query(
  .data,
  selected_columns = attr(.data, "selected_columns"),
  filtered_rows = attr(.data, "filtered_rows")
)
}
\arguments{
\item{.data}{Either a data.frame, Arrow Table, substrait_dplyr_query object; whatever}

\item{selected_columns}{Columns to select}

\item{filtered_rows}{Rows to filter

This will need refining later but it's a useful convenience function for now
as it can operate on whatever}
}
\description{
Create a substrait_dplyr_query object
}
