% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-arrow.R
\name{substrait_eval_arrow}
\alias{substrait_eval_arrow}
\title{Evaluate a Plan using Arrow}
\usage{
substrait_eval_arrow(plan, tables)
}
\arguments{
\item{plan}{A substrait.Plan}

\item{tables}{A named list of data frames, Datasets, or Tables
corresponding to NamedTable objects.}
}
\value{
An \link[arrow:Table]{arrow::Table}.
}
\description{
Evaluate a Plan using Arrow
}
