% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-rel.R
\name{substrait_fetch}
\alias{substrait_fetch}
\title{Append a Substrait Fetch Relation}
\usage{
substrait_fetch(.compiler, offset, count)
}
\arguments{
\item{.compiler}{A \code{\link[=substrait_compiler]{substrait_compiler()}} or object that can be coerced to one}

\item{...}{Expressions}
}
\value{
A modified \code{.compiler}
}
\description{
Append a Substrait Fetch Relation
}
\examples{
substrait_fetch(
  duckdb_substrait_compiler(data.frame(x = 1:10)),
  offset = 2,
  count = 2
)

}
