% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-rel.R
\name{substrait_filter}
\alias{substrait_filter}
\title{Append a Substrait Project Relation}
\usage{
substrait_filter(.compiler, ...)
}
\arguments{
\item{.compiler}{A \code{\link[=substrait_compiler]{substrait_compiler()}} or object that can be coerced to one}

\item{...}{Filter expressions}
}
\value{
A modified \code{.compiler}
}
\description{
Append a Substrait Project Relation
}
\examples{
substrait_filter(
  duckdb_substrait_compiler(data.frame(a = 1, b = "one")),
  a > 0
)

}
