% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{substrait_schema.data.frame}
\alias{substrait_schema.data.frame}
\title{Create a substrait schema}
\usage{
\method{substrait_schema}{data.frame}(x)
}
\arguments{
\item{x}{data.frame object}
}
\description{
Create a substrait schema
}
