% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{substrait_schema}
\alias{substrait_schema}
\title{Create a substrait schema}
\usage{
substrait_schema(x)
}
\arguments{
\item{x}{object to be turned into a schema}
}
\description{
Create a substrait schema
}
