% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort-rel.R
\name{substrait_sort}
\alias{substrait_sort}
\alias{substrait_sort_field}
\title{Append a Substrait Project Relation}
\usage{
substrait_sort(.builder, ...)

substrait_sort_field(expr, direction = "SORT_DIRECTION_UNSPECIFIED")
}
\arguments{
\item{.builder}{A \code{\link[=substrait_builder]{substrait_builder()}} or object that can be coerced to one}

\item{...}{Expressions that evaluate to an Expression or SortField}

\item{expr}{An expression that evaluates to an Expression}

\item{direction}{A SortField.SortDirection}
}
\value{
A modified \code{.builder}
}
\description{
Append a Substrait Project Relation
}
\examples{
substrait_sort(
  data.frame(a = 1, b = "one"),
  a,
  substrait_sort_field(b, "SORT_DIRECTION_DESC_NULLS_LAST")
)

}
