% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{substrait_tbl}
\alias{substrait_tbl}
\title{Build a Substrait plan}
\usage{
substrait_tbl(tbl, ..., compiler = substrait_compiler())
}
\arguments{
\item{tbl}{A table-like object that can be converted to a ReadRel.}

\item{...}{Passed to S3 methods.}

\item{compiler}{a \code{\link[=substrait_compiler]{substrait_compiler()}}}
}
\value{
An object of class 'substrait_builder'
}
\description{
Build a Substrait plan
}
\examples{
substrait_tbl(data.frame(col1 = 1 , col2 = "one"))

}
