% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrait_dplyr_query.R
\name{build_filters}
\alias{build_filters}
\title{Take filtered rows and create the appropriate substrait message}
\usage{
build_filters(input, filters)
}
\arguments{
\item{input}{input table}

\item{filters}{list of quosures}
}
\description{
Take filtered rows and create the appropriate substrait message
}
