% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-filters.R
\name{build_filters}
\alias{build_filters}
\title{Take filtered rows and create the appropriate substrait message}
\usage{
build_filters(filters, .data)
}
\arguments{
\item{filters}{list of quosures}
}
\description{
Take filtered rows and create the appropriate substrait message
}
