% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-duckdb.R
\name{duckdb_get_substrait}
\alias{duckdb_get_substrait}
\alias{duckdb_from_substrait}
\alias{has_duckdb_with_substrait}
\title{DuckDB Substrait Interface}
\usage{
duckdb_get_substrait(sql, tables = list())

duckdb_from_substrait(
  plan,
  tables = list(),
  col_names = plan$relations[[1]]$root$names,
  as_data_frame = TRUE
)

has_duckdb_with_substrait()
}
\arguments{
\item{sql}{An SQL expression from which to generate a Substrait plan}

\item{tables}{A named list of tables to populate the database}

\item{plan}{A substrait.Plan proto object}

\item{col_names}{The final column names for the result}

\item{as_data_frame}{Use \code{FALSE} to return an \link[arrow:Table]{arrow::Table} instead of a
data.frame.}
}
\value{
\itemize{
\item \code{duckdb_get_substrait()}: a substrait.Plan protobuf object
\item \code{duckdb_from_substrait()}: A data.frame or arrow Table
}
}
\description{
DuckDB Substrait Interface
}
\examples{
\dontshow{if (has_duckdb_with_substrait()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
plan <- duckdb_get_substrait(
  "SELECT * from mtcars WHERE mpg > 30",
  tables = list(mtcars = mtcars)
)

duckdb_from_substrait(plan, tables = list(mtcars = mtcars))
\dontshow{\}) # examplesIf}
}
