% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 6 columns.
}
\usage{
data(example_data)
}
\description{
Data for use in examples and tests with multiple different data types
}
\examples{
data(example_data)
}
\keyword{datasets}
