% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-filter.R
\name{filter.substrait_dplyr_query}
\alias{filter.substrait_dplyr_query}
\title{Filter column}
\usage{
\method{filter}{substrait_dplyr_query}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{substrait_dplyr_query object or data.frame}

\item{...}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Expressions that return a
logical value, and are defined in terms of the variables in \code{.data}.
If multiple expressions are included, they are combined with the \code{&} operator.
Only rows for which all conditions evaluate to \code{TRUE} are kept.}

\item{.preserve}{Relevant when the \code{.data} input is grouped.
If \code{.preserve = FALSE} (the default), the grouping structure
is recalculated based on the resulting data, otherwise the grouping is kept as is.}
}
\description{
Filter column
}
