% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{substrait_builder}
\alias{substrait_builder}
\alias{substrait_builder.substrait_builder}
\alias{substrait_builder.default}
\title{Build a Substrait plan}
\usage{
substrait_builder(rel, ..., consumer = GenericConsumer$new())

\method{substrait_builder}{substrait_builder}(rel, ..., consumer = GenericConsumer$new())

\method{substrait_builder}{default}(rel, ..., consumer = GenericConsumer$new())
}
\arguments{
\item{consumer}{A \link{Consumer} instance}
}
\value{
An object of class 'substrait_builder'
}
\description{
Build a Substrait plan
}
\examples{
substrait_builder(data.frame(col1 = 1 , col2 = "one"))

}
