% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{substrait_builder}
\alias{substrait_builder}
\alias{substrait_builder.substrait_builder}
\alias{substrait_builder.default}
\alias{substrait_evaluate}
\title{Build a Substrait plan}
\usage{
substrait_builder(rel, ..., consumer = GenericConsumer$new())

\method{substrait_builder}{substrait_builder}(rel, ..., consumer = GenericConsumer$new())

\method{substrait_builder}{default}(rel, ..., consumer = GenericConsumer$new())

substrait_evaluate(builder, ...)
}
\arguments{
\item{rel}{A table-like object with which to create a builder.}

\item{...}{Passed to the \link{GenericConsumer} when creating a new builder}

\item{consumer}{A \link{GenericConsumer} instance}

\item{builder}{A \code{\link[=substrait_builder]{substrait_builder()}}}
}
\value{
An object of class 'substrait_builder'
}
\description{
Build a Substrait plan
}
\examples{
substrait_builder(data.frame(col1 = 1 , col2 = "one"))

}
