% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{substrait_compiler}
\alias{substrait_compiler}
\alias{substrait_compiler.SubstraitCompiler}
\alias{substrait_compiler.default}
\title{Build a Substrait plan}
\usage{
substrait_compiler(rel, ..., consumer = SubstraitCompiler$new())

\method{substrait_compiler}{SubstraitCompiler}(rel, ..., consumer = SubstraitCompiler$new())

\method{substrait_compiler}{default}(rel, ..., consumer = SubstraitCompiler$new())
}
\arguments{
\item{rel}{A table-like object with which to create a builder.}

\item{...}{Passed to the \link{SubstraitCompiler} when creating a new builder}

\item{consumer}{A \link{SubstraitCompiler} instance}

\item{builder}{A \code{\link[=substrait_compiler]{substrait_compiler()}}}
}
\value{
An object of class 'substrait_compiler'
}
\description{
Build a Substrait plan
}
\examples{
substrait_compiler(data.frame(col1 = 1 , col2 = "one"))

}
