% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{substrait_compiler}
\alias{substrait_compiler}
\alias{substrait_compiler.SubstraitCompiler}
\alias{substrait_compiler.default}
\title{Initialize a Substrait Compiler}
\usage{
substrait_compiler(object, ...)

\method{substrait_compiler}{SubstraitCompiler}(object, ...)

\method{substrait_compiler}{default}(object, ...)
}
\arguments{
\item{object}{A table-like object with which to create a compiler.}

\item{...}{Passed to the \link{SubstraitCompiler} when creating a new compiler}
}
\value{
An object of class 'substrait_compiler'
}
\description{
Creates a \link{SubstraitCompiler} instance initialized with \code{object}
(e.g., a \code{data.frame()}).
}
\examples{
substrait_compiler(data.frame(col1 = 1 , col2 = "one"))

}
