% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrait_dplyr_query.R
\name{substrait_dplyr_query}
\alias{substrait_dplyr_query}
\title{Create a Substrait query}
\usage{
substrait_dplyr_query(
  .data,
  selected_columns = attr(.data, "selected_columns"),
  filtered_rows = attr(.data, "filtered_rows"),
<<<<<<< HEAD
  arrange_vars = attr(.data, "arrange_vars"),
  arrange_desc = attr(.data, "arrange_desc")
=======
  arrange_vars = attr(.data, "arrange_vars")
>>>>>>> Add dplyr::arrange and sort field
)
}
\arguments{
\item{.data}{Either a data.frame, Arrow Table, substrait_dplyr_query object; whatever}

\item{selected_columns}{Columns to select}

\item{filtered_rows}{Rows to filter}

\item{arrange_vars}{Columns to sort on}

\item{arrange_desc}{Are sort columns in descending order?}
}
\value{
An object of class 'substrait_dplyr_query'
}
\description{
Create a Substrait query
}
\examples{
substrait_dplyr_query(mtcars)
}
