% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-rel.R
\name{substrait_filter}
\alias{substrait_filter}
\title{Append a Substrait Project Relation}
\usage{
substrait_filter(.builder, ...)
}
\arguments{
\item{.builder}{A \code{\link[=substrait_builder]{substrait_builder()}} or object that can be coerced to one}

\item{...}{Filter expressions}
}
\value{
A modified \code{.builder}
}
\description{
Append a Substrait Project Relation
}
\examples{
substrait_filter(
  data.frame(a = 1, b = "one"),
  c = a + 1
)

}
