% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-rel.R
\name{substrait_project}
\alias{substrait_project}
\title{Append a Substrait Project Relation}
\usage{
substrait_project(.compiler, ...)
}
\arguments{
\item{.compiler}{A \code{\link[=substrait_compiler]{substrait_compiler()}} or object that can be coerced to one}

\item{...}{Expressions}
}
\value{
A modified \code{.compiler}
}
\description{
Append a Substrait Project Relation
}
\examples{
substrait_project(
  data.frame(a = 1, b = "one"),
  c = a + 1
)

}
